requirements.txt
# a2wsgi==1.10.4
# anyio==4.3.0
# asgiref==3.7.2
# attrs==23.2.0
# blinker==1.7.0
# certifi==2024.2.2
# charset-normalizer==3.3.2
# click==8.1.7
# connexion==3.0.6
# exceptiongroup==1.2.0
# Flask==3.0.2
# h11==0.14.0
# httpcore==1.0.4
# httptools==0.6.1
# httpx==0.27.0
# idna==3.6
# importlib_metadata==7.0.2
# inflection==0.5.1
# itsdangerous==2.1.2
# Jinja2==3.1.3
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
# MarkupSafe==2.1.5
# python-dotenv==1.0.1
# python-multipart==0.0.9
# PyYAML==6.0.1
# referencing==0.34.0
# requests==2.31.0
# rpds-py==0.18.0
# sniffio==1.3.1
# starlette==0.37.2
# swagger_ui_bundle==1.1.0
# typing_extensions==4.10.0
# urllib3==2.2.1
# uvicorn==0.29.0
# uvloop==0.19.0
# watchfiles==0.21.0
# websockets==12.0
# Werkzeug==3.0.1
# zipp==3.18.1


aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
Flask==1.1.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
swagger-ui-bundle==0.0.9
urllib3==1.26.7
Werkzeug==1.0.1